+++ CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG
 - Update pip/setuptools/wheel in virtual environment before installing dependencies
   [#5042](https://github.com/pulumi/pulumi/pull/5042)
 
+- Add ability to change a secrets provider for the current stack
+  [#5031](https://github.com/pulumi/pulumi/pull/5031)
+
 ## 2.7.1 (2020-07-22)
 
 - Fix logic to parse pulumi venv on github action